waiting for db info to test new profile jsp

+++ src/main/java/com/codeup/adlister/dao/ListAdsDao.java
@@ -22,7 +22,7 @@ public class ListAdsDao implements Ads {
         }
         // we'll assign an "id" here based on the size of the ads list
         // really the dao would handle this
-        ad.setId((long) ads.size());
+        ad.setId(ads.size());
         ads.add(ad);
         return ad.getId();
     }
